requirements.txt
torch
torchvision
numpy
pillow
jupyter
ipykernel
matplotlib
tqdm